Realized training procedure needed complete redo.
Training cannot evaluate the success of the network, that has
to be done by the application. Only the application knows what
the heck it's trying to learn.

On a different note, I added some more debug and realized this
stinkin' thing doesn't work at all. The results appear to be
entirely random. Any "successful" learnings were coincidence.