Fundamental flaw in my understanding of neural net
I had a fundamental misunderstanding. I was
calculating the activations incorrectly. The
correct activation level involves a dot product
of the weights and activations. I had written
something similar, but not a proper dot product.
I had wires crossed, in a sense. So I changed
the representation (basically swapped breadth
and depth) to more simply see proper matrix
multiplication.

I also split up the back propagation into discrete
steps of:
1. Find error in output
2. Find error in hidden
3. Apply change to output weights
4. Apply change to hidden weights

I made that change based on Wikipedia's algorithm
and the fact that I was having so much trouble
following the flow. I was getting confused.

Anyway, final story is that it works for the XOR
gate, and this time I mean it!