Wrote methods and tests for saving/loading weights.
The methods store the network's weights in a specified
file. On subsuquent uses the network can load the saved
weights.

There are also tests written, but they aren't completely
thorough -- there are situations in which an improperly
formatted file could go uncaught and get loaded. I'm not
too worried about this; it isn't my focus.